+++ src/space/perceptual.rs
@@ -2,6 +2,8 @@
 mod okhsl;
 #[cfg(feature = "space-okhsv")]
 mod okhsv;
+#[cfg(feature = "space-okhwb")]
+mod okhwb;
 #[cfg(feature = "space-oklab")]
 mod oklab;
 #[cfg(feature = "space-oklch")]
@@ -11,6 +13,8 @@ mod oklch;
 pub use okhsl::Okhsl;
 #[cfg(feature = "space-okhsv")]
 pub use okhsv::Okhsv;
+#[cfg(feature = "space-okhwb")]
+pub use okhwb::Okhwb;
 #[cfg(feature = "space-oklab")]
 pub use oklab::Oklab;
 #[cfg(feature = "space-oklch")]