+++ src/space/physiological/lms.rs
@@ -23,6 +23,8 @@ use crate::space::Luv;
 use crate::space::Okhsl;
 #[cfg(feature = "space-okhsv")]
 use crate::space::Okhsv;
+#[cfg(feature = "space-okhwb")]
+use crate::space::Okhwb;
 #[cfg(feature = "space-oklab")]
 use crate::space::Oklab;
 #[cfg(feature = "space-oklch")]
@@ -571,6 +573,13 @@ impl From<Okhsl> for Lms {
 impl From<Okhsv> for Lms {
   fn from(okhsv: Okhsv) -> Self {
     okhsv.to_lms()
+  }
+}
+
+#[cfg(feature = "space-okhwb")]
+impl From<Okhwb> for Lms {
+  fn from(okhwb: Okhwb) -> Self {
+    okhwb.to_lms()
   }
 }